Bugfix #16 - Add support for --allow-subset-of-names
There will be situations where certificates vouch for domains that are not
actually used nor implemented. The subset parameter allows the validation
process to proceed if at least one of the existing domains passes the
challenge. This will implicitly fix the issue encountered where Entrust
certificates grant an extra www based domain which we do not actually have
routing for. We are setting this to be a default true parameter in order
to make sure certbot is more likely to succeed in default scenarios.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>