Mass update readme documentation
This commit focuses on updating, standardizing and ordering all parameters
related to deploying and managing Certbot. By ensuring that all parameters
use the same naming scheme, it will reduce confusion on intent. As well,
we have changed the default cronjob schedule to run bi-weekly on Monday
and Thursday as opposed to running every 12 hours. Since certificates tend
to last a while, running it more frequently than twice a week is a waste
of cluster resources.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>